requirements.txt
boto3==1.9.214
botocore==1.12.214
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-cors-headers==3.0.2
django-filter==2.2.0
django-oauth-toolkit==1.2.0
django-storages==1.7.1
djangorestframework==3.10.1
docutils==0.15.2
httplib2==0.13.1
idna==2.8
jmespath==0.9.4
Markdown==3.1.1
mysqlclient==1.4.4
oauth2==1.9.0.post1
oauthlib==3.0.2
Pillow==6.1.0
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
rest-condition==1.0.3
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
